pyproject.toml
[tool.poetry]
name = "gestaltconfig"
version = "0.1.0"
description = ""
authors = ["Frederick.Ollinger <follinge@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.3"
pyaml = "^24.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
